01-Run Llama 2 on Google Colab/requirements.txt
llama-cpp-python
huggingface_hub
streamlit